backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.36
alembic==1.14.0
pyjwt==2.10.1
bcrypt==4.2.1
python-dotenv==1.0.1
